requirements.txt
matplotlib~=3.3.4
numpy~=1.19.2
Pillow~=8.2.0
torch
torchvision~=0.9.1
tensorboard
future
tqdm~=4.60.0

scikit-image~=0.18.1
opencv-python
pandas~=1.2.4